Change the version number to long

+++ chat2db-server/chat2db-server-start/src/main/java/ai/chat2db/server/start/config/config/AutomaticUpgrade.java
@@ -112,7 +112,7 @@ public class AutomaticUpgrade {
         String[] versionArray = upgrade.getVersion().split("\\.");
         String[] localVersionArray = localVersion.split("\\.");
         for (int i = 0; i < versionArray.length; i++) {
-            if (Integer.parseInt(versionArray[i]) > Integer.parseInt(localVersionArray[i])) {
+            if (Long.parseLong(versionArray[i]) > Long.parseLong(localVersionArray[i])) {
                 return true;
             }
         }
